Discord API is super limited
You are not able to do most of these interactions through Oauth and need to setup a bot etc which is more effort than I would have liked. Will revisit this in the future.

This branch is not in a stable state but its far enough along for anyone to pick up.